feat: Merge O3DBroadcast into Open3DStream plugin as a module (Issue #28)
Consolidate O3DBroadcast functionality into the Open3DStream plugin as an
integrated module, simplifying distribution and installation.

CHANGES:
- Move Open3DBroadcast code from separate plugin to module within Open3DStream
  * New module path: plugins/unreal/Open3DStream/Source/Open3DBroadcast/
  * Both modules share unified .uplugin descriptor

- Implement selective module compilation via O3DS_WITH_BROADCAST flag
  * Default: both modules compiled
  * Can be disabled with -DO3DS_WITH_BROADCAST=0
  * Added BuildConfig.txt documentation

- Delete obsolete standalone O3DBroadcast plugin
  * Removed: plugins/unreal/Open3DBroadcast/
  * Cleaned up all references to separate plugin

- Update CI workflows for unified plugin
  * Remove separate build steps for O3DBroadcast
  * Simplified build matrix and artifact packaging
  * Updated release notes and installation instructions

- Update documentation
  * Clean user-facing documentation (no references to merged plugins)
  * Updated copilot-instructions.md for module structure
  * Added plugin README with module descriptions
  * Updated Build scripts documentation

BENEFITS:
- Single plugin installation for users
- Cleaner distribution (one .uplugin file)
- Proper module dependencies enforced
- Build flexibility via flags
- Simpler CI/CD workflows

TESTING:
- Both modules compile independently
- Plugin dependency chain validated
- No breaking changes to receiver functionality
- Build configuration tested
